Simplify the InputField/InputFieldComponent, rename them
The Input Field is now Field Handler because it manages the underlying
Input Field. Input Field is now Field Component, it's a better name.

Remove not needed getter methods and move them to the Abstract Field Handler
as protected methods, because they are only needed internal.

Move Abstract Field Handler to the public module because the purpose
of the class is to have a default ctor for creating new fields.