Add reflections KDE 4 icons
This icon theme is a compilation of a few icon themes and a few sparse
icons found on www.deviantart.com and www.kde-look.org. It is by no
means complete but hopefully (in time) it will be. The theme is mostly
based on the reflection icon set by styrizo
(http://styrizo.deviantart.com) and has been enhanced with the minimal
perception set and allblack by mandarancio
(http://mandarancio.deviantart.com/) and the ecqlipse icon set by
http://chrfb.deviantart.com/.

The script used to convert the icons is based on the one for
Ekisho_Project
(http://www.kde-look.org/content/show.php/Ekisho_Project?content=55588).
The muon icon is a modification of the muon icon by user JOnas
(http://kde-look.org/content/show.php/Muon+Icon?content=145610). Thanks
to user sfranzen for providing an svg version of the lyx icon! The juk
icon is a modification from Bespin.

The above creative authors have produced and keep doing an amazing work
so I wish to thank them and express my gratitude for granting me their
permission to create and share this compilation
theme. Thank you!